requirements.txt
aiofiles
aiohttp
aiosignal
altair
annotated-types
anyio
asttokens
async-timeout
attrs
backcall
backoff
beautifulsoup4
cachetools
certifi
chardet
charset-normalizer
chromadb
click
clickhouse-connect
coloredlogs
comm
contourpy
cycler
dataclasses
dataclasses-json
debugpy
decorator
duckdb
entrypoints
exceptiongroup
executing
fastapi
ffmpy
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
gradio
gradio_client
greenlet
h11
hnswlib
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib-metadata
importlib-resources
ipykernel
ipython
jedi
jieba
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
langchain
langsmith
linkify-it-py
lxml
loguru
lz4
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdit-py-plugins
mdurl
monotonic
mpmath
multidict
mypy-extensions
nest-asyncio
nltk
numexpr
numpy
onnxruntime
openai
orjson
overrides
packaging
pandas
parso
pexpect
pickleshare
Pillow
pip
platformdirs
posthog
prompt-toolkit
protobuf
psutil
ptyprocess
pulsar-client
pure-eval
pydantic
pydantic_core
pydub
Pygments
PyJWT
PyMuPDF
PyMuPDFb
pyparsing
python-dateutil
python-dotenv
python-magic
python-multipart
pytz
PyYAML
pyzmq
referencing
regex
requests
rouge-chinese
rpds-py
scikit-learn
scipy
semantic-version
setuptools
six
sniffio
SQLAlchemy
stack-data
starlette
sympy
tabulate
tenacity
threadpoolctl
tokenizers
toolz
tornado
tqdm
traitlets
typing_extensions
typing-inspect
tzdata
uc-micro-py
unstructured
urllib3
uvicorn
uvloop
watchfiles
wcwidth
websocket-client
websockets
wheel
yarl
zhipuai
zipp
zstandard
langchain-community
transformers
vllm
modelscope
sentence-transformers
langchain-chroma